requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-dotenv==1.4.2
django-filter==23.3
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
mysqlclient==2.2.0
Naked==0.1.32
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
status==0.2.5
tzdata==2023.3
urllib3==2.0.6
